move file_content_digest to file_utils
